rename to error-overlay-call-stack
